% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChunks.R
\name{foldChunks}
\alias{foldChunks}
\title{Fold code chunks in an r notebook file.  Output remains visible.}
\usage{
foldChunks(currentChunkOnly = F)
}
\arguments{
\item{currentChunkOnly}{Fold the current chunk, or all of them?}
}
\description{
Fold code chunks in an r notebook file.  Output remains visible.
}
\examples{
\dontrun{
foldCode(T) # fold only the code region in the selection
foldCode() # fold all code regions
}
}
