% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldCode.R
\name{foldCode}
\alias{foldCode}
\title{Fold code regions in an r notebook file}
\usage{
foldCode(currentChunkOnly = F)
}
\arguments{
\item{currentChunkOnly}{Fold the current chunk, or all of them?}
}
\description{
Fold code regions in an r notebook file
}
\examples{
\dontrun{
foldCode(T) # fold only the code region in the selection
foldCode() # fold all code regions
}
}
