% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChunks.R
\name{foldCurrentChunk}
\alias{foldCurrentChunk}
\title{Fold active code chunk in an r notebook file}
\usage{
foldCurrentChunk()
}
\description{
Fold active code chunk in an r notebook file
}
\examples{
\dontrun{
foldCurrentCode() # fold only the chunk in the selection
}
}
